feat: fonctionne; readme update
initial commit
