--- mongo-init.js
@@ -1,15 +1,0 @@
-db = db.getSiblingDB('trends');
-
-db.createUser(
-    {
-        user: "trending",
-        pwd: "trending",
-        roles: [
-            {
-                role: "readWrite",
-                db: "trends"
-            }
-        ]
-    }
-);
-db.createCollection('trends');
